refactor(geometry): Consolidate mazes/ subdirectory into graph/
Networks and mazes are all graphs. The graph/ subdirectory is the
canonical location for all graph-based geometries.

Changes:
- Migrated 17 import statements from geometry.mazes to geometry.graph
- Deleted mazes/ subdirectory (9 files)
- Updated code files (4), tests (2), docs (2), examples (4)

All 46 maze-related tests pass.

Related to Issue #320 (geometry subdirectory cleanup)
