+++ examples/advanced/capacity_constrained_mfg/capacity_field.py
@@ -61,7 +61,7 @@ class CapacityField:
         bounds: Bounding box ((x_min, x_max), (y_min, y_max))
 
     Examples:
-        >>> from mfg_pde.geometry.mazes import create_perfect_maze
+        >>> from mfg_pde.geometry.graph import create_perfect_maze
         >>> maze = create_perfect_maze(rows=20, cols=20, wall_thickness=3)
         >>> maze_array = maze.to_numpy_array(wall_thickness=3)
         >>> capacity = CapacityField.from_maze_geometry(maze_array, wall_thickness=3)
@@ -377,7 +377,7 @@ def visualize_capacity_field(
         figsize: Figure size (width, height)
 
     Examples:
-        >>> from mfg_pde.geometry.mazes import create_perfect_maze
+        >>> from mfg_pde.geometry.graph import create_perfect_maze
         >>> maze = create_perfect_maze(rows=20, cols=20, wall_thickness=3)
         >>> maze_array = maze.to_numpy_array(wall_thickness=3)
         >>> capacity = CapacityField.from_maze_geometry(maze_array)
