+++ tests/unit/test_voronoi_maze.py
@@ -69,7 +69,7 @@ class TestUnionFind:
 
     def test_initialization(self):
         """Test UnionFind initialization."""
-        from mfg_pde.geometry.mazes.voronoi_maze import UnionFind
+        from mfg_pde.geometry.graph.voronoi_maze import UnionFind
 
         uf = UnionFind(5)
 
@@ -79,7 +79,7 @@ class TestUnionFind:
 
     def test_union_operation(self):
         """Test union of two sets."""
-        from mfg_pde.geometry.mazes.voronoi_maze import UnionFind
+        from mfg_pde.geometry.graph.voronoi_maze import UnionFind
 
         uf = UnionFind(5)
 
@@ -92,7 +92,7 @@ class TestUnionFind:
 
     def test_union_already_connected(self):
         """Test union of already connected elements."""
-        from mfg_pde.geometry.mazes.voronoi_maze import UnionFind
+        from mfg_pde.geometry.graph.voronoi_maze import UnionFind
 
         uf = UnionFind(5)
 
@@ -103,7 +103,7 @@ class TestUnionFind:
 
     def test_path_compression(self):
         """Test path compression optimization."""
-        from mfg_pde.geometry.mazes.voronoi_maze import UnionFind
+        from mfg_pde.geometry.graph.voronoi_maze import UnionFind
 
         uf = UnionFind(5)
 
@@ -237,7 +237,7 @@ class TestVoronoiMazeGenerator:
 
     def test_spanning_tree_connectivity(self):
         """Test that spanning tree connects all points."""
-        from mfg_pde.geometry.mazes.voronoi_maze import UnionFind
+        from mfg_pde.geometry.graph.voronoi_maze import UnionFind
 
         config = VoronoiMazeConfig(rows=40, cols=40, num_points=10, seed=42)
         generator = VoronoiMazeGenerator(config)
